cleaned some unnecessary header files

+++ lib/vector/vector.h
@@ -4,17 +4,17 @@ union Vector
 {
 	struct
 	{
-		float angle, strength;
+		int angle, strength;
 	};
 
 	struct 
 	{
-		float x, y;
+		int x, y;
 	};
 
 	Vector()
 		: x(0), y(0) {}
 
-	Vector(float x, float y)
+	Vector(int x, int y)
 		: x(x), y(y) {}
 };
